chore(shared): introduce nondominium_shared crate, eliminate type duplication
Rename crates/utils -> crates/shared (package nondominium_shared).
Make hdk optional behind `coordinator` feature so integrity zomes and
native Sweettest crates can import shared types directly.

Moved to crates/shared/src/types.rs:
- LifecycleStage, PropertyRegime, ResourceNature (were duplicated between
  zome_resource_integrity and zome_lobby_integrity — lobby had an incorrect
  ADR comment claiming 'WASM forbids cross-crate type imports between DNAs')
- VfAction, NdoLinkType, BeneficiaryRef, BenefitType, BenefitClause

Moved to crates/shared/src/io/{lobby,governance}.rs:
- AnnounceNdoInput, LobbyAgentProfileInput, UpdateNdoAnnouncementInput,
  GroupDescriptorStub, CreateAgreementInput, UpdateAgreementInput,
  ValidateContributionInput, CreateNdoHardLinkInput, GetNdoHardLinksByTypeInput

Updated all coordinator Cargo.tomls (path + coordinator feature), all
integrity zome Cargo.tomls, both Sweettest Cargo.tomls, workspace Cargo.toml.
All nondominium_utils imports renamed to nondominium_shared.

Sweettest mirror structs in governance/mod.rs and lobby/mod.rs replaced
with direct imports; VfAction/LifecycleStage/PropertyRegime/ResourceNature
String literals replaced with typed enum values.

cargo check --target wasm32-unknown-unknown (with RUSTFLAGS getrandom_backend=custom): Finished
