+++ Cargo.toml
@@ -14,7 +14,7 @@ members = [
     "dnas/nondominium/zomes/integrity/zome_resource",
     "dnas/nondominium/zomes/coordinator/zome_gouvernance",
     "dnas/nondominium/zomes/integrity/zome_gouvernance",
-    "crates/utils",
+    "crates/shared",
     "dnas/nondominium/tests",
     "dnas/lobby/tests",
 ]
@@ -32,13 +32,14 @@ default-members = [
     "dnas/nondominium/zomes/coordinator/zome_gouvernance",
     "dnas/nondominium/zomes/integrity/zome_gouvernance",
     "dnas/nondominium/zomes/coordinator/misc",
-    "crates/utils",
+    "crates/shared",
     "dnas/lobby/zomes/coordinator/zome_lobby_coordinator",
     "dnas/lobby/zomes/integrity/zome_lobby_integrity",
 ]
 resolver = "2"
 
 [workspace.dependencies]
+nondominium_shared = { path = "crates/shared" }
 hdi = "^0.7.0"
 hdk = "^0.6.0"
 holochain_serialized_bytes = "*"
