+++ crates/shared/src/errors.rs
