+++ crates/shared/src/paths.rs
