--- crates/utils/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "nondominium_utils"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-hdk = { workspace = true }
-serde = { workspace = true }
-thiserror = { workspace = true }
