+++ dnas/nondominium/zomes/coordinator/zome_gouvernance/Cargo.toml
@@ -14,4 +14,4 @@ holochain_serialized_bytes = { workspace = true }
 thiserror = { workspace = true }
 blake3 = "1.5"
 zome_gouvernance_integrity = { path = "../../integrity/zome_gouvernance" }
-nondominium_utils = { path = "../../../../../crates/utils" }
+nondominium_shared = { workspace = true, features = ["coordinator"] }
