+++ dnas/nondominium/zomes/coordinator/zome_person/Cargo.toml
@@ -13,4 +13,4 @@ serde = { workspace = true }
 holochain_serialized_bytes = { workspace = true }
 thiserror = { workspace = true }
 zome_person_integrity = { path = "../../integrity/zome_person" }
-nondominium_utils = { path = "../../../../../crates/utils" }
+nondominium_shared = { workspace = true, features = ["coordinator"] }
